[Rgen] Move most of the naming logic to the Nomenclator.
This way we have all the naming in a single location and it is easier to
make sure that we do name variables in a consistent manner.

This also removes the need of some of the special code for the generator
in the data model.
