+++ src/rgen/Microsoft.Macios.Generator/DataModel/TypeInfo.Generator.cs
@@ -12,12 +12,7 @@ readonly partial struct TypeInfo {
 	/// True if the type needs to use a stret call.
 	/// </summary>
 	public bool NeedsStret { get; init; }
-
-	/// <summary>
-	/// Get the name of the variable for the type when it is used as a return value.
-	/// </summary>
-	public string ReturnVariableName => "ret"; // nothing fancy for now
-
+	
 	internal TypeInfo (ITypeSymbol symbol, Compilation compilation) : this (symbol)
 	{
 		IsNativeEnum = symbol.HasAttribute (AttributesNames.NativeEnumAttribute);
